+++ core/src/com/ray3k/skincomposer/dialog/DialogDrawables.java
@@ -193,7 +193,8 @@ public class DialogDrawables extends Dialog {
                 atlas = null;
             }
             if (!main.getAtlasData().atlasCurrent) {
-                main.getAtlasData().writeAtlas(Gdx.files.internal("atlas-internal-settings.json"));
+                FileHandle defaultsFile = Main.appFolder.child("texturepacker/atlas-internal-settings.json");
+                main.getAtlasData().writeAtlas(defaultsFile);
                 main.getAtlasData().atlasCurrent = true;
             }
             atlas = main.getAtlasData().getAtlas();
@@ -2227,7 +2228,8 @@ public class DialogDrawables extends Dialog {
         
         try {
             if (!main.getAtlasData().atlasCurrent) {
-                main.getAtlasData().writeAtlas(Gdx.files.internal("atlas-internal-settings.json"));
+                FileHandle defaultsFile = Main.appFolder.child("texturepacker/atlas-internal-settings.json");
+                main.getAtlasData().writeAtlas(defaultsFile);
                 main.getAtlasData().atlasCurrent = true;
             }
         } catch (Exception e) {