+++ CHANGES.md
@@ -8,6 +8,7 @@
 * Added dialog to fix the minWidth and minHeight settings of drawables when opening old projects.
 * Added tooltips when hovering over items in the preview for custom classes.
 * Added zoom option to preview.
+* Added option to change preview texture packer settings. See Project >> Settings.
 * Fixed nine patch editor not showing drawables when selected from Content in Preview area.
 * Added wrapping to Style Property entries to prevent overflow.
 * Fixed character filtering in Spinner so that negative values can be entered properly.